+++ dee/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - dee: add metadata for future beacons
+- dee: inspect tlock age header
 
 ## [0.0.11] - 2023-08-08